Code documentation

SeedMixFunctions.jl is the included package that provides the mix evaluation and GA run initiation functions. Use of these functions isn't required to run an analysis - the file multistart.jl in the scripts folder (documented here) is the main way to run the analysis - but the functions are documented here for comprehensiveness.

Data loading function

SeedMixFunctions.SpeciesData — Method

SpeciesData(tablefile::String, phylofile::String)

Reads data from tablefile and phylofile to construct the SpeciesData structure.

Data types

These are the structs that you need to run these analyses.

SeedMixFunctions.SpeciesData — Type

SpeciesData struct

SeedMixFunctions.MixData — Type

MixData struct

SeedMixFunctions.MixRequirements — Type

Structure capturing requirements for each mix:

nspecs::Int64
totalweight::Float64
minweight::Float64
[maxcost::Float64]

Note there is an a three argument constructor that leaves out maxcost. It will set maxcost to 0.0, which is treated as no limit in the MOEA.

GA functions

Functions to run or post-process runs of the GA

SeedMixFunctions.RunParams — Type

Struct that defines parameters for the genetic algorithm (GA) function:

nstarts::Int64  -  number of separate GA iterations to run
popsize::Int64  -  population size for each GA run
ngens::Int64    -  number of generations to simulate in each GA run

SeedMixFunctions.multi_start — Function

Perform a multi-start run of the GA

SeedMixFunctions.do_run — Function

Perform a single run of the GA with given data and parameters

SeedMixFunctions.save_results — Function

Writes output files for genomes, objectives, and mix details

SeedMixFunctions.pool_spec_selections — Function

Performs union of species found in results' mixes

Evaluation functions

These are functions that evaluate a mix for a specific metric.

SeedMixFunctions.get_phylo_dist — Method

get_phylo_dist(mix::MixData, sd::SpeciesData)

Computes summed pairwise phylogenetic distance.

SeedMixFunctions.get_shannon — Function

Calculates Shannon Diversity

SeedMixFunctions.get_cost — Function

Calculates cost of the mix: sum of weight * cost

SeedMixFunctions.get_consval — Function

Calculates conservatism value

SeedMixFunctions.get_bloom — Function

Calculates bloom index: sum of square root of # flowering species per month

SeedMixFunctions.get_grass_spec_frac — Function

Calculates the fraction of mix species that are grasses

SeedMixFunctions.get_grass_weight_frac — Function

Calculates the weight fraction of seeds that are grasses

Index

SeedMixFunctions.MixData

SeedMixFunctions.MixRequirements

SeedMixFunctions.RunParams

SeedMixFunctions.SpeciesData

SeedMixFunctions.SpeciesData

SeedMixFunctions.do_run

SeedMixFunctions.get_bloom

SeedMixFunctions.get_consval

SeedMixFunctions.get_cost

SeedMixFunctions.get_grass_spec_frac

SeedMixFunctions.get_grass_weight_frac

SeedMixFunctions.get_phylo_dist

SeedMixFunctions.get_shannon

SeedMixFunctions.multi_start

SeedMixFunctions.pool_spec_selections

SeedMixFunctions.save_results